load spct to table

+++ src/main/java/org/project1/nhom8/dto/SPCTViewModel.java
@@ -33,5 +33,5 @@ public class SPCTViewModel {
     private String trangThai;
 
     @DataHeader(name = "giá")
-    private Float gia;
+    private double gia;
 }
